+++ lib/src/enums/app_state.dart
@@ -6,6 +6,7 @@ enum AppState {
   /// [background] means that the app is in the background.
   background,
 
-  /// [terminated] means that the app is completely closed.
+  /// [terminated] means that the app is terminated i.e. not running
+  /// in the background or foreground.
   terminated,
 }