+++ lib/src/models/notification_on_tap_details.dart
@@ -1,11 +1,14 @@
 import 'package:firebase_notifications_handler/src/enums/app_state.dart';
 
 class NotificationOnTapDetails {
-  NotificationOnTapDetails({
+  const NotificationOnTapDetails({
     required this.appState,
     required this.payload,
   });
 
-  final Map<dynamic, dynamic> payload;
+  /// The app state when the notification was tapped.
   final AppState appState;
+
+  /// The payload of the notification.
+  final Map<String, dynamic> payload;
 }